Combine smart-goals and home routes
